+++ CHANGELOG.md
@@ -5,9 +5,11 @@
 ### Changed
 
 - Include chart name when installing with a custom release name ([#227], [#228]).
+- Orphaned resources are deleted ([#255]).
 
 [#227]: https://github.com/stackabletech/superset-operator/pull/227
 [#228]: https://github.com/stackabletech/superset-operator/pull/228
+[#255]: https://github.com/stackabletech/superset-operator/pull/255
 
 ## [0.5.0] - 2022-06-30